refactor: move public specific components into public site folder p4
