+++ sites/public/src/lib/applications/configInterfaces.ts
