+++ sites/public/src/pages/applications/start/autofill.tsx
@@ -19,7 +19,7 @@ import { useForm } from "react-hook-form"
 import FormsLayout from "../../../layouts/forms"
 import { useFormConductor } from "../../../lib/hooks"
 import FormSummaryDetails from "../../../components/forms/applications/FormSummaryDetails"
-import AutofillCleaner from "../../../lib/appAutofill"
+import AutofillCleaner from "../../../lib/applications/appAutofill"
 import { useRouter } from "next/router"
 import { UserStatus } from "../../../lib/constants"
